removes local storage from web-client
It was used in the past to cope with page refresh,
when the app would try to retake control
over an already existing player.
This can't have worked for a long time,
since a refresh will also issue a player delete.

Removing this feature also prepares replacing the player manager
by the players Vuex module